requirements.txt
attrs==21.4.0
bottle==0.12.18
certifi==2021.10.8
charset-normalizer==2.0.11
docopt==0.6.2
ExifRead==2.3.1
filetype==1.0.10
idna==3.3
iniconfig==1.1.1
mysql-connector==2.2.9
packaging==21.3
parsing==2.0.0
Paste==3.4.4
pipreqs==0.4.11
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytz==2021.3
requests==2.27.1
sh==1.14.0
six==1.16.0
toml==0.10.2
urllib3==1.26.8
yarg==0.1.9

